requirements.txt
fastapi
uvicorn[standard]
requests
python-multipart
redis
google-generativeai
Pillow
